[PM-32427] Reorganize types
